Vastly improved logging functions
using the new log methods, logging is now much easier.
